requirements.txt
pyspark
boto3
psycopg2
pyyaml
kafka-python
